Changed data type from int to Integer in all Model classes.  Fixed dependencies.  Removed all not null constraints from the database.  The not null constraints should be done in Model classes using annotations.

+++ src/main/java/edu/yu/einstein/wasp/batch/WorkflowPoller.java
@@ -58,7 +58,7 @@ public class WorkflowPoller {
 
 			// should be the same so just get first, 
 			// TODO check for npe
-			if (state.getStatejob().get(0).getJob().getWorkflowId() != workflow.getWorkflowId()) {
+			if (state.getStatejob().get(0).getJob().getWorkflowId().intValue() != workflow.getWorkflowId().intValue()) {
 				continue;
 			}
